Source/requirements.txt
altgraph==0.17.3
appdirs==1.4.4
atomicwrites==1.4.0
autopep8==2.0.4
backports.weakref==1.0.post1
bkcharts==0.2
brotlipy==0.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
comtypes==1.1.10
constantly==15.1.0
cssselect==1.1.0
cssutils==2.9.0
CTkListbox==1.1
customtkinter==5.2.1
darkdetect==0.8.0
distlib==0.3.7
et-xmlfile==1.1.0
filelock==3.13.1
fonttools==4.25.0
idna==2.10
inflection==0.5.1
Jinja2==3.0.1
MarkupSafe==2.1.3
numpy==1.26.2
packaging==23.2
pefile==2023.2.7
Pillow==10.1.0
pipenv==2023.11.15
platformdirs==4.0.0
pycodestyle==2.11.1
pycparser==2.21
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.10
pytz==2023.3.post1
pywin32-ctypes==0.2.2
requests==2.31.0
setuptools==69.0.2
six==1.16.0
urllib3==2.1.0
virtualenv==20.25.0
